откатил обратно на js чтобы сделать счетчик
