+++ render-game.js
