+++ tsconfig.json
@@ -1,12 +1,12 @@
-{
-    "CompilerOptions": {
-        "strict": true,
-        "outDir": "./dist/",
-        "noImplicitAny": true,
-        "module": "es6",
-        "target": "es6",
-        "allowJs": true,
-        "sourceMap": true,
-        "moduleResolution": "node"
-    }
-}
+// {
+//     "CompilerOptions": {
+//         "strict": true,
+//         "outDir": "./dist/",
+//         "noImplicitAny": true,
+//         "module": "es6",
+//         "target": "es6",
+//         "allowJs": true,
+//         "sourceMap": true,
+//         "moduleResolution": "node"
+//     }
+// }
